sources/requirements.txt
fastapi[all]
typer[all]
uvicorn
subprocrunner
loguru
pyyaml
